refactor: implement useMergeQueries hook and streamline vault queries
- Introduced a new `useMergeQueries` hook to simplify the merging of query results, enhancing code modularity and readability.
- Updated `RemoteStateDependant` to utilize the new hook, replacing the previous `mergeQueries` utility.
- Removed obsolete `mergeQueries` utility function, streamlining the codebase.
- Refactored multiple components to use the new `useVaults` and `useVaultNames` hooks for improved state management.
