+++ clients/desktop/src/state/RemoteStateDependant.tsx
@@ -1,10 +1,10 @@
+import { VaultsProvider } from '@core/ui/vault/state/vaults'
 import { ChildrenProp } from '@lib/ui/props'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
-import { mergeQueries } from '@lib/ui/query/utils/mergeQueries'
 import { StrictText } from '@lib/ui/text'
-import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 
+import { useMergeQueries } from '../lib/hooks/useMergeQueries'
 import { Center } from '../lib/ui/layout/Center'
 import { ProductLogoBlock } from '../ui/logo/ProductLogoBlock'
 import { useVaultsQuery } from '../vault/queries/useVaultsQuery'
@@ -14,17 +14,19 @@ export const RemoteStateDependant = ({ children }: ChildrenProp) => {
   const vaults = useVaultsQuery()
   const vaultFolders = useVaultFoldersQuery()
 
-  const query = useMemo(
-    () => mergeQueries({ vaults, vaultFolders }),
-    [vaultFolders, vaults]
-  )
+  const query = useMergeQueries({
+    vaults,
+    vaultFolders,
+  })
 
   const { t } = useTranslation()
 
   return (
     <MatchQuery
       value={query}
-      success={() => children}
+      success={({ vaults }) => (
+        <VaultsProvider value={vaults}>{children}</VaultsProvider>
+      )}
       error={() => (
         <Center>
           <StrictText>{t('failed_to_load')}</StrictText>
