+++ clients/desktop/src/vault/components/ActiveVaultGuard.tsx
@@ -1,9 +1,9 @@
+import { useVaults } from '@core/ui/vault/state/vaults'
 import { getVaultId } from '@core/ui/vault/Vault'
 import { ChildrenProp } from '@lib/ui/props'
 import { useEffect } from 'react'
 
 import { useAppNavigate } from '../../navigation/hooks/useAppNavigate'
-import { useVaults } from '../queries/useVaultsQuery'
 import { CurrentVaultProvider } from '../state/currentVault'
 import { CurrentVaultCoinsProvider } from '../state/currentVaultCoins'
 import { useCurrentVaultId } from '../state/currentVaultId'
