+++ clients/desktop/src/vault/components/EmptyVaultsOnly.tsx
@@ -1,9 +1,9 @@
+import { useVaults } from '@core/ui/vault/state/vaults'
 import { ChildrenProp } from '@lib/ui/props'
 import { isEmpty } from '@lib/utils/array/isEmpty'
 import { useEffect } from 'react'
 
 import { useAppNavigate } from '../../navigation/hooks/useAppNavigate'
-import { useVaults } from '../queries/useVaultsQuery'
 
 export const EmptyVaultsOnly = ({ children }: ChildrenProp) => {
   const navigate = useAppNavigate()
