--- clients/desktop/src/vault/hooks/useVaultNames.ts
@@ -1,10 +1,0 @@
-import { withoutDuplicates } from '@lib/utils/array/withoutDuplicates'
-import { useMemo } from 'react'
-
-import { useVaults } from '../queries/useVaultsQuery'
-
-export const useVaultNames = () => {
-  const vaults = useVaults()
-
-  return useMemo(() => withoutDuplicates(vaults.map(v => v.name)), [vaults])
-}
