+++ clients/desktop/src/vault/keygen/join/JoinKeygenVaultProvider.tsx
@@ -7,13 +7,13 @@ import {
 import { CurrentHexChainCodeProvider } from '@core/ui/mpc/state/currentHexChainCode'
 import { useMpcDevice } from '@core/ui/mpc/state/mpcDevice'
 import { MpcLocalPartyIdProvider } from '@core/ui/mpc/state/mpcLocalPartyId'
+import { useVaults } from '@core/ui/vault/state/vaults'
 import { ChildrenProp } from '@lib/ui/props'
 import { matchRecordUnion } from '@lib/utils/matchRecordUnion'
 import { pick } from '@lib/utils/record/pick'
 import { useMemo } from 'react'
 
 import { useAppPathState } from '../../../navigation/hooks/useAppPathState'
-import { useVaults } from '../../queries/useVaultsQuery'
 
 export const JoinKeygenVaultProvider: React.FC<ChildrenProp> = ({
   children,
