+++ clients/desktop/src/vault/keygen/shared/mutations/useDklsKeygen.ts
@@ -16,6 +16,7 @@ import { useMpcLocalPartyId } from '@core/ui/mpc/state/mpcLocalPartyId'
 import { useMpcPeers } from '@core/ui/mpc/state/mpcPeers'
 import { useMpcServerUrl } from '@core/ui/mpc/state/mpcServerUrl'
 import { useMpcSessionId } from '@core/ui/mpc/state/mpcSession'
+import { useVaults } from '@core/ui/vault/state/vaults'
 import { Vault } from '@core/ui/vault/Vault'
 import { match } from '@lib/utils/match'
 import { getLastItemOrder } from '@lib/utils/order/getLastItemOrder'
@@ -26,7 +27,6 @@ import {
   GetLocalUIEcdsa,
   GetLocalUIEdDSA,
 } from '../../../../../wailsjs/go/tss/TssService'
-import { useVaults } from '../../../queries/useVaultsQuery'
 import { KeygenResolver } from './KeygenResolver'
 
 export const useDklsKeygen = (): KeygenResolver => {
