+++ clients/desktop/src/vault/keysign/join/KeysignVaultGuard.tsx
@@ -1,4 +1,5 @@
 import { MpcLocalPartyIdProvider } from '@core/ui/mpc/state/mpcLocalPartyId'
+import { useVaults } from '@core/ui/vault/state/vaults'
 import { getVaultId } from '@core/ui/vault/Vault'
 import { Button } from '@lib/ui/buttons/Button'
 import { ChildrenProp } from '@lib/ui/props'
@@ -8,7 +9,6 @@ import { Link } from 'react-router-dom'
 import { makeAppPath } from '../../../navigation'
 import { useAppPathState } from '../../../navigation/hooks/useAppPathState'
 import { FullPageFlowErrorState } from '../../../ui/flow/FullPageFlowErrorState'
-import { useVaults } from '../../queries/useVaultsQuery'
 import { CurrentVaultProvider } from '../../state/currentVault'
 import { CurrentVaultCoinsProvider } from '../../state/currentVaultCoins'
 import { useCurrentVaultId } from '../../state/currentVaultId'
