+++ clients/desktop/src/vault/setup/fast/SetupFastVaultPage.tsx
@@ -1,3 +1,4 @@
+import { CreateVaultFlowProviders } from '@core/ui/mpc/keygen/create/CreateVaultFlowProviders'
 import { VaultSecurityTypeProvider } from '@core/ui/mpc/keygen/create/state/vaultSecurityType'
 import { Match } from '@lib/ui/base/Match'
 import { useStepNavigation } from '@lib/ui/hooks/useStepNavigation'
@@ -12,7 +13,6 @@ import { PasswordProvider } from '../../server/password/state/password'
 import { ServerPasswordHintStep } from '../../server/password-hint/ServerPasswordHintStep'
 import { PasswordHintProvider } from '../../server/password-hint/state/password-hint'
 import { SetupVaultNameStep } from '../SetupVaultNameStep'
-import { CreateVaultFlowProviders } from '../shared/CreateVaultFlowProviders'
 import { SetupVaultServerStep } from './SetupVaultServerStep'
 
 const steps = [
