+++ clients/desktop/src/vaults/components/Vaults.tsx
@@ -1,7 +1,7 @@
+import { useFolderlessVaults } from '@core/ui/vault/state/vaults'
 import { getVaultId } from '@core/ui/vault/Vault'
 import { NonEmptyOnly } from '@lib/ui/base/NonEmptyOnly'
 
-import { useFolderlessVaults } from '../../vault/queries/useVaultsQuery'
 import { CurrentVaultProvider } from '../../vault/state/currentVault'
 import { VaultListItem } from './VaultListItem'
 import { VaultsContainer } from './VaultsContainer'
