+++ clients/desktop/src/vaults/folder/VaultFolderPage.tsx
@@ -1,3 +1,4 @@
+import { useFolderVaults } from '@core/ui/vault/state/vaults'
 import { getVaultId } from '@core/ui/vault/Vault'
 import { EditIcon } from '@lib/ui/icons/EditIcon'
 import { VStack } from '@lib/ui/layout/Stack'
@@ -8,7 +9,6 @@ import { PageHeader } from '../../ui/page/PageHeader'
 import { PageHeaderBackButton } from '../../ui/page/PageHeaderBackButton'
 import { PageHeaderIconButton } from '../../ui/page/PageHeaderIconButton'
 import { PageHeaderTitle } from '../../ui/page/PageHeaderTitle'
-import { useFolderVaults } from '../../vault/queries/useVaultsQuery'
 import { CurrentVaultProvider } from '../../vault/state/currentVault'
 import { VaultListItem } from '../components/VaultListItem'
 import { useCurrentVaultFolder } from './state/currentVaultFolder'
