+++ clients/desktop/src/vaults/folder/manage/ManageFolderVaults.tsx
@@ -1,3 +1,4 @@
+import { useFolderVaults } from '@core/ui/vault/state/vaults'
 import { getVaultId } from '@core/ui/vault/Vault'
 import { VStack } from '@lib/ui/layout/Stack'
 import { Text } from '@lib/ui/text'
@@ -13,7 +14,6 @@ import {
   DnDItemHighlight,
 } from '../../../lib/ui/list/item/DnDItemContainer'
 import { useUpdateVaultOrderMutation } from '../../../vault/mutations/useUpdateVaultOrderMutation'
-import { useFolderVaults } from '../../../vault/queries/useVaultsQuery'
 import { CurrentVaultProvider } from '../../../vault/state/currentVault'
 import { FolderVaultOption } from '../addVaults/FolderVaultOption'
 import { useRemoveVaultFromFolderMutation } from '../mutations/useRemoveVaultFromFolderMutation'
