+++ clients/desktop/src/vaults/folders/create/CreateVaultFolderPage.tsx
@@ -1,3 +1,4 @@
+import { useFolderlessVaults } from '@core/ui/vault/state/vaults'
 import { NonEmptyOnly } from '@lib/ui/base/NonEmptyOnly'
 import { Button } from '@lib/ui/buttons/Button'
 import { VStack } from '@lib/ui/layout/Stack'
@@ -11,7 +12,6 @@ import { useNavigateBack } from '../../../navigation/hooks/useNavigationBack'
 import { FlowPageHeader } from '../../../ui/flow/FlowPageHeader'
 import { PageContent } from '../../../ui/page/PageContent'
 import { PageFooter } from '../../../ui/page/PageFooter'
-import { useFolderlessVaults } from '../../../vault/queries/useVaultsQuery'
 import { FolderVaultsInput } from '../../folder/addVaults/FolderVaultsInput'
 import { useCreateVaultFolderMutation } from '../../folder/mutations/useCreateVaultFolderMutation'
 import { useVaultFolders } from '../queries/useVaultFoldersQuery'
