+++ clients/desktop/src/vaults/manage/AddVaultsToFolder.tsx
@@ -1,7 +1,7 @@
+import { useFolderlessVaults } from '@core/ui/vault/state/vaults'
 import { getVaultId } from '@core/ui/vault/Vault'
 import { isEmpty } from '@lib/utils/array/isEmpty'
 
-import { useFolderlessVaults } from '../../vault/queries/useVaultsQuery'
 import { CurrentVaultProvider } from '../../vault/state/currentVault'
 import { AddVaultsToFolderContainer } from '../folder/addVaults/AddVaultsToFolderContainer'
 import { FolderVaultOption } from '../folder/addVaults/FolderVaultOption'
