+++ core/ui/mpc/keygen/create/state/vaultName.tsx
@@ -4,21 +4,17 @@ import { withoutUndefined } from '@lib/utils/array/withoutUndefined'
 import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 
+import { useVaultNames } from '../../../../vault/state/vaults'
 import { useVaultSecurityType } from './vaultSecurityType'
 
 export const { useState: useVaultName, provider: VaultNameProvider } =
   getStateProviderSetup<string>('VaultName')
 
-type GeneratedVaultNameProviderProps = ChildrenProp & {
-  existingNames: string[]
-}
-
-export const GeneratedVaultNameProvider = ({
-  children,
-  existingNames,
-}: GeneratedVaultNameProviderProps) => {
+export const GeneratedVaultNameProvider = ({ children }: ChildrenProp) => {
   const vaultSecurityType = useVaultSecurityType()
   const { t } = useTranslation()
+
+  const existingNames = useVaultNames()
 
   const intitialValue = useMemo(() => {
     const prefix = `${t(vaultSecurityType)} ${t('vault')} #`
